+++ src/__mocks__/@formkit/auto-animate/react.js
@@ -1,0 +1,2 @@
+// Github issue related to this mock: https://github.com/formkit/auto-animate/issues/29
+export const useAutoAnimate = () => [null];