+++ src/components/ExternalLink/ExternalLink.tsx
@@ -4,7 +4,7 @@ import { FC } from 'react';
 
 const ExternalLink: FC<LinkProps> = (linkProps) => (
   <>
-    <Link {...linkProps} target="_blank" rel="noreferrer">
+    <Link color="info.main" {...linkProps} target="_blank" rel="noreferrer">
       {linkProps.children}
       <LaunchIcon fontSize={'small'} sx={{ fontSize: '0.8rem', ml: '1px', verticalAlign: 'top' }} />
     </Link>