+++ src/components/forms/completions/DefaultCompletionsForm/DefaultCompletionsForm.tsx
@@ -1,4 +1,5 @@
 import styled from '@emotion/styled';
+import { useAutoAnimate } from '@formkit/auto-animate/react';
 import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline';
 import DeleteIcon from '@mui/icons-material/Delete';
 import HelpIcon from '@mui/icons-material/Help';
@@ -16,11 +17,10 @@ import {
   Typography,
   useTheme
 } from '@mui/material';
-import { FC, useRef, useState } from 'react';
+import { FC, useState } from 'react';
 import { Controller, useFieldArray, useFormContext } from 'react-hook-form';
 
 import { useAppSelector } from '../../../../hooks/redux';
-import { useAutoAnimate } from '../../../../hooks/useAutoAnimate';
 import Completion from '../../../../models/Completion';
 import ThemeMode from '../../../../models/ThemeMode';
 import { selectThemeMode } from '../../../../store/theme/theme.selectors';
@@ -56,8 +56,7 @@ const DefaultCompletionsForm: FC<Props> = ({
   });
   const themeMode = useAppSelector(selectThemeMode);
   const theme = useTheme();
-  const animateRef = useRef(null);
-  useAutoAnimate(animateRef);
+  const [animateRef] = useAutoAnimate();
 
   const fieldsetSx = {
     bgcolor: themeMode === ThemeMode.LIGHT ? theme.palette.grey[100] : theme.palette.grey[900],
@@ -76,7 +75,9 @@ const DefaultCompletionsForm: FC<Props> = ({
   const id = open ? 'popover-' + name : undefined;
 
   return (
-    <Box ref={animateRef}>
+    // Little hack on Box ref to make it work with TS: https://github.com/mui/material-ui/issues/17010#issuecomment-615577360
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    <Box {...({ ref: animateRef } as any)}>
       {fields.map((field, index) => (
         <FieldsetGrid columns={2} container key={field.id} sx={fieldsetSx} {...gridProps}>
           <Grid item xs={2} sx={{ display: 'flex', justifyContent: 'space-between', mb: 1 }}>