+++ src/config/Constants.ts
@@ -1,6 +1,6 @@
 import { subDays } from 'date-fns';
 
-export const DRAWER_MENU_WIDTH = '250px';
+export const DRAWER_MENU_WIDTH = '260px';
 export const JSON_DATA_EXTENSION = '.sv.json'; // SuiVie json
 export const PAYPAL_DONATE_URL =
   'https://www.paypal.com/donate/?business=UF2GH7ZPMLES8&no_recurring=0&item_name=Chaque+don+me+permet+de+continuer+%C3%A0+d%C3%A9velopper+et+%C3%A0+am%C3%A9liorer+l%27application+SuiVie%2C+merci+%21+%3A%29&currency_code=EUR';