+++ src/pages/About/PackageVersion.tsx
@@ -1,13 +1,10 @@
-import { Box, Typography } from '@mui/material';
+import { Box, BoxProps, Typography } from '@mui/material';
+import { FC } from 'react';
 
 import packageInfo from '../../../package.json';
 
-const PackageVersion = () => (
-  <Box
-    bgcolor="primary.main"
-    color="common.white"
-    px={1}
-    style={{ bottom: 0, right: 0, position: 'fixed' }}>
+const PackageVersion: FC<BoxProps> = (boxProps) => (
+  <Box bgcolor="primary.main" color="common.white" px={1} {...boxProps}>
     <Typography>v{packageInfo.version}</Typography>
   </Box>
 );